added refactor to consultation cancellation
