+++ src/main/kotlin/com/jsrdev/med_api/domain/consult/Consult.kt
@@ -1,5 +1,6 @@
 package com.jsrdev.med_api.domain.consult
 
+import com.jsrdev.med_api.domain.consult.cancel.CancellationReason
 import com.jsrdev.med_api.domain.patient.Patient
 import com.jsrdev.med_api.domain.physician.Physician
 import jakarta.persistence.*
