+++ src/main/kotlin/com/jsrdev/med_api/domain/consult/cancel/CancellationReason.kt
@@ -1,4 +1,4 @@
-package com.jsrdev.med_api.domain.consult
+package com.jsrdev.med_api.domain.consult.cancel
 
 enum class CancellationReason {
     PATIENT_DESISTED,
